Remove semicolon from end on lines in php.ini
php.ini files should not have a semicolon at the end on a line.

Can
cause issues in some hosting configurations where the php.ini file is
not processed directly by PHP. An example is cPanel with its MultiPHP
INI Editor enabled. It will include the semicolon with the value leading
to "charset `UTF-8;' not supported" errors.