Bootstrap 5.2.1 fix tooltips (another attempt)
https://github.com/opencart/opencart/pull/11759

No JS errors now. Tooltips are shown every time they are hovered.

Checked in Firefox and Chrome.

Sometimes, when you slide very fast between multiple buttons, a tooltip looses it's style and position. But at least this is not so noticeable as current implementation when tooltips are always displaced on second hover.
This looks like a BS tooltip or popover bug.