Removing subscriptions in editOrder() method
As per this issue commit's explanation: https://github.com/opencart/opencart/issues/11935 . The charge() method in the cron/subscription needs to create a new order by extension prior to return the subscription_status_id. If active, the addTransaction() gets added based on this (which we already have in the cron/subscription). Therefore, to ensure that a new order gets created, we need to ensure the subscription is removed from the editOrder().